Work around state persisting between scenarios
Work around the feature that all scenarios in a given file use the same webdriver session, persisting application state between scenarios. It's done to speed up testing and can be worked around by splitting scenarios into separate files, but that couples the specification structure to the technical limitations of the testing tools, so work around it with a tag that optionally triggers creation of a new session at the end of scenarios.